+++ src/rpc/progress_api.rs
@@ -5,12 +5,12 @@
 
 use std::sync::atomic;
 
-use crate::ipld::{
-    ProgressBarCurrentTotalPair, WALK_SNAPSHOT_PROGRESS_DB_GC, WALK_SNAPSHOT_PROGRESS_EXPORT,
-};
+use crate::db::parity_db::DATABASE_DUMP_PROGRESS;
+use crate::ipld::{WALK_SNAPSHOT_PROGRESS_DB_GC, WALK_SNAPSHOT_PROGRESS_EXPORT};
 use crate::rpc_api::progress_api::{GetProgressParams, GetProgressResult, GetProgressType};
 
 use crate::rpc::*;
+use crate::utils::io::progress_bar::ProgressBarCurrentTotalPair;
 
 pub(in crate::rpc) async fn get_progress(
     Params((typ,)): Params<GetProgressParams>,
@@ -18,6 +18,7 @@ pub(in crate::rpc) async fn get_progress(
     let tracker: &ProgressBarCurrentTotalPair = match typ {
         GetProgressType::SnapshotExport => &WALK_SNAPSHOT_PROGRESS_EXPORT,
         GetProgressType::DatabaseGarbageCollection => &WALK_SNAPSHOT_PROGRESS_DB_GC,
+        GetProgressType::DatabaseDump => &DATABASE_DUMP_PROGRESS,
     };
 
     Ok((